requirements.txt
python-dotenv>=1.0.0
keyring>=25.0.0
sounddevice>=0.4.6
numpy>=1.24.0
pynput>=1.7.6
pystray>=0.19.5
Pillow>=10.0.0
pyperclip>=1.8.2
httpx>=0.25.0
